pipeline/cricket/requirements.txt
beautifulsoup4
openai
psycopg2-binary
pytz
requests